+++ CHANGES.md
@@ -1,6 +1,8 @@
 # django-pg-zero-downtime-migrations changelog
 
-## 0.6
+## 0.6 
+  - added postgres 12 safe `NOT NULL` constraint creation
+  - added safe `NOT NULL` constraint creation for extra permissions for `pg_catalog.pg_attribute` with `ZERO_DOWNTIME_MIGRATIONS_USE_NOT_NULL=USE_PG_ATTRIBUTE_UPDATE_FOR_SUPERUSER` option enabled
   - marked `AddField` with `null=False` parameter and compatible `CHECK IS NOT NULL` constraint option as unsafe operation and avoid `ZERO_DOWNTIME_MIGRATIONS_USE_NOT_NULL` value in this case
   - added version to package
   - fixed pypi README images links